requirements.txt
alpaca-py==0.30.1
websockets==12.0
aiohttp==3.9.5
pytest==8.2.0
python-dotenv==1.0.1
pyfiglet==1.0.2
readchar==4.0.5
matplotlib==3.9.0
